refactor(edge-groups): replace function with class
