+++ app/edge/views/groups/edgeGroupsView.html
